+++ R/dbn_visualization.R
@@ -38,7 +38,7 @@ node_levels <- function(net, order, lvl = 1, acc = NULL){
 #' plot_network(fit) # Works for both the structure and the fitted net
 #' }
 #' @export
-plot_network <- function(structure){
+plot_static_network <- function(structure){
   check_opt_pkgs_available()
   initial_bn_check(structure)
   if(is_dbn_or_dbnfit(structure))
@@ -211,4 +211,19 @@ plot_dynamic_network <- function(structure, offset = 200, subset_nodes = NULL){
                            nodesIdSelection = T)
 
   return(ret)
+}
+
+#' Plots a Bayesian networks or a dynamic Bayesian network
+#'
+#' Selects the appropriate function depending on the class of the object
+#' @param structure the structure or fit of the network.
+#' @param ... additional parameters for the visualization of a DBN
+#' @export
+plot_network <- function(structure, ...){
+  initial_bn_or_dbn_check(structure)
+  
+  if(is_dbn_or_dbnfit(structure))
+    plot_dynamic_network(structure, ...)
+  else
+    plot_static_network(structure)
 }